requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
asgiref==3.4.1
async-timeout==4.0.1
attrs==19.3.0
Automat==0.8.0
base58==2.1.1
bitarray==1.2.2
blinker==1.4
certifi==2019.11.28
chardet==3.0.4
charset-normalizer==2.0.7
Click==7.0
cloud-init==21.1
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==2.8
cytoolz==0.11.2
dbus-python==1.2.16
distro==1.4.0
distro-info===0.23ubuntu1
Django==3.2.9
django-cors-headers==3.10.0
django-filter==21.1
djangorestframework==3.12.4
entrypoints==0.3
eth-abi==2.1.1
eth-account==0.5.6
eth-hash==0.3.2
eth-keyfile==0.5.1
eth-keys==0.3.3
eth-rlp==0.2.1
eth-typing==2.2.2
eth-utils==1.10.0
frozenlist==1.2.0
gunicorn==20.1.0
hexbytes==0.2.2
httplib2==0.14.0
hyperlink==19.0.0
idna==2.8
importlib-metadata==4.8.2
incremental==16.10.1
ipfshttpclient==0.8.0a2
Jinja2==2.10.1
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lru-dict==1.1.7
Markdown==3.3.6
MarkupSafe==1.1.0
more-itertools==4.2.0
multiaddr==0.0.9
multidict==5.2.0
netaddr==0.8.0
netifaces==0.10.4
oauthlib==3.1.0
parsimonious==0.8.1
pexpect==4.6.0
protobuf==3.19.1
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycryptodome==3.11.0
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyrsistent==0.15.5
pyserial==3.4
python-apt==2.0.0+ubuntu0.20.4.4
python-debian===0.1.36ubuntu1
python-decouple==3.5
pytz==2021.3
PyYAML==5.3.1
requests==2.22.0
requests-unixsocket==0.2.0
rlp==2.0.1
SecretStorage==2.3.1
service-identity==18.1.0
simplejson==3.16.0
six==1.14.0
sos==4.1
sqlparse==0.4.2
ssh-import-id==5.10
systemd-python==234
toolz==0.11.2
Twisted==18.9.0
typing-extensions==4.0.0
ubuntu-advantage-tools==27.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
varint==1.0.2
wadllib==1.3.3
web3==5.25.0
websockets==9.1
yarl==1.7.2
zipp==3.6.0
zope.interface==4.7.1
